requirements.txt
asyncpg
loguru
python_dateutil
dotenv
